[clang][deps] Consolidate types into new `DependencyGraph.h`
